backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
gitpython>=3.1.43
anthropic>=0.28.0
python-dotenv>=1.0.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
httpx>=0.27.0
pypdf>=4.0.0
pyyaml>=6.0.0
